Check reorganization for errors
