+++ src/components/Cart/Cart.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import "./Cart.css";
 
 const Cart = () => {
