+++ src/components/ListingView/ListingView.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import './ListingView.css';
