+++ src/components/NotFound/NotFound.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Link } from 'react-router-dom'
 import './NotFound.css';
