+++ src/services/firebase/listings/index.js
@@ -1,3 +1,5 @@
-import { createListing, getAllListings, getListing, getListings } from './listings';
+import { createListing, getAllListings, getListing, getListings, Listing } from './listings';
+import Item from './item';
+import Property from './property';
 
-export { createListing, getAllListings, getListing, getListings };
+export { createListing, getAllListings, getListing, getListings, Listing, Item, Property };
